requirements.txt
discord
discord-py-interactions
python-decouple
oauth2client
gspread
google-api-python-client